feat(types): migrate async-queue to typescript
